Generell opprydding i tester
