+++ packages/api/Tests/UnitTests/InvoiceRate/InvoiceRateServiceTest.cs
@@ -20,6 +20,7 @@ public class InvoiceRateServiceTest
     private readonly IOptionsMonitor<TimeEntryOptions> _options;
     private readonly IRedDaysService _redDaysService;
     private readonly Mock<IUserContext> _userContextMock;
+    
     public InvoiceRateServiceTest()
     {
         var entryOptions = new TimeEntryOptions
@@ -73,7 +74,7 @@ public class InvoiceRateServiceTest
 
         var service = CreateInvoiceRateService(hours);
 
-        decimal invoiceRate = await service.GetEmployeeInvoiceRateForPeriod(new DateTime(2022, 01, 03), new DateTime(2022, 01, 03));
+        var invoiceRate = await service.GetEmployeeInvoiceRateForPeriod(new DateTime(2022, 01, 03), new DateTime(2022, 01, 03));
 
         Assert.Equal(0.8m, invoiceRate);
     }
@@ -143,7 +144,7 @@ public class InvoiceRateServiceTest
 
         var service = CreateInvoiceRateService(hours);
 
-        decimal invoiceRate = await service.GetEmployeeInvoiceRateForPeriod(new DateTime(2022, 4, 11), new DateTime(2022, 4, 17));
+        var invoiceRate = await service.GetEmployeeInvoiceRateForPeriod(new DateTime(2022, 4, 11), new DateTime(2022, 4, 17));
 
         Assert.Equal(0.6m, invoiceRate);
     }
@@ -212,7 +213,7 @@ public class InvoiceRateServiceTest
 
         var service = CreateInvoiceRateService(hours);
 
-        decimal invoiceRate = await service.GetEmployeeInvoiceRateForPeriod(new DateTime(2022, 12, 12), new DateTime(2022, 12, 18));
+        var invoiceRate = await service.GetEmployeeInvoiceRateForPeriod(new DateTime(2022, 12, 12), new DateTime(2022, 12, 18));
 
         Assert.Equal(1m, invoiceRate);
     }
@@ -225,7 +226,7 @@ public class InvoiceRateServiceTest
 
         var hours = new List<Hours>();
 
-        for (int i = 0; i < 59; i++)
+        for (var i = 0; i < 59; i++)
         {
             var newDate = startDate.AddDays(i);
             if (newDate.DayOfWeek == DayOfWeek.Sunday || newDate.DayOfWeek == DayOfWeek.Saturday)
@@ -250,13 +251,13 @@ public class InvoiceRateServiceTest
 
         var statistics = await service.GetEmployeeInvoiceStatisticsByPeriod(startDate, endDate, InvoicePeriods.Daily, ExtendPeriod.None, false);
 
-        Assert.Equal(21, statistics.BillableHours.Count());
-        Assert.Equal(21, statistics.NonBillableHours.Count());
-        Assert.Equal(21, statistics.VacationHours.Count());
-        Assert.Equal(21, statistics.InvoiceRate.Count());
-        Assert.Equal(21, statistics.NonBillableInvoiceRate.Count());
-        Assert.Equal(21, statistics.Start.Count());
-        Assert.Equal(21, statistics.End.Count());
+        Assert.Equal(21, statistics.BillableHours.Length);
+        Assert.Equal(21, statistics.NonBillableHours.Length);
+        Assert.Equal(21, statistics.VacationHours.Length);
+        Assert.Equal(21, statistics.InvoiceRate.Length);
+        Assert.Equal(21, statistics.NonBillableInvoiceRate.Length);
+        Assert.Equal(21, statistics.Start.Length);
+        Assert.Equal(21, statistics.End.Length);
 
         Assert.Equal(7.5m, statistics.BillableHours[0]);
         Assert.Equal(0m, statistics.VacationHours[0]);
@@ -283,7 +284,7 @@ public class InvoiceRateServiceTest
 
         var hours = new List<Hours>();
 
-        for (int i = 0; i < 59; i++)
+        for (var i = 0; i < 59; i++)
         {
             var newDate = startDate.AddDays(i);
             if (newDate.DayOfWeek == DayOfWeek.Sunday || newDate.DayOfWeek == DayOfWeek.Saturday)
@@ -340,7 +341,7 @@ public class InvoiceRateServiceTest
         var endDate = new DateTime(2022, 1, 9);
 
         var hours = new List<Hours>();
-        for (int i = 0; i < 7; i++)
+        for (var i = 0; i < 7; i++)
         {
             var newDate = startDate.AddDays(i);
             hours.Add(new Hours
@@ -385,7 +386,7 @@ public class InvoiceRateServiceTest
         for (int i = 0; i < 14; i++)
         {
             var newDate = startDate.AddDays(i);
-            if (newDate.DayOfWeek == DayOfWeek.Sunday || newDate.DayOfWeek == DayOfWeek.Saturday)
+            if (newDate.DayOfWeek is DayOfWeek.Sunday or DayOfWeek.Saturday)
             {
                 continue;
             }
@@ -428,10 +429,10 @@ public class InvoiceRateServiceTest
         var endDate = new DateTime(2022, 12, 31);
 
         var hours = new List<Hours>();
-        for (int i = 0; i < 365; i++)
+        for (var i = 0; i < 365; i++)
         {
             var newDate = startDate.AddDays(i);
-            if (newDate.DayOfWeek == DayOfWeek.Sunday || newDate.DayOfWeek == DayOfWeek.Saturday)
+            if (newDate.DayOfWeek is DayOfWeek.Sunday or DayOfWeek.Saturday)
             {
                 continue;
             }
@@ -472,10 +473,10 @@ public class InvoiceRateServiceTest
         var endDate = new DateTime(2022, 2, 15);
 
         var hours = new List<Hours>();
-        for (int i = 0; i < 31; i++)
+        for (var i = 0; i < 31; i++)
         {
             var newDate = startDate.AddDays(i);
-            if (newDate.DayOfWeek == DayOfWeek.Sunday || newDate.DayOfWeek == DayOfWeek.Saturday)
+            if (newDate.DayOfWeek is DayOfWeek.Sunday or DayOfWeek.Saturday)
             {
                 continue;
             }
@@ -496,13 +497,13 @@ public class InvoiceRateServiceTest
 
         var statistics = await service.GetEmployeeInvoiceStatisticsByPeriod(startDate, endDate, InvoicePeriods.Monthly, ExtendPeriod.None, true);
 
-        Assert.Equal(2, statistics.BillableHours.Count());
-        Assert.Equal(2, statistics.NonBillableHours.Count());
-        Assert.Equal(2, statistics.VacationHours.Count());
-        Assert.Equal(2, statistics.InvoiceRate.Count());
-        Assert.Equal(2, statistics.NonBillableInvoiceRate.Count());
-        Assert.Equal(2, statistics.Start.Count());
-        Assert.Equal(2, statistics.End.Count());
+        Assert.Equal(2, statistics.BillableHours.Length);
+        Assert.Equal(2, statistics.NonBillableHours.Length);
+        Assert.Equal(2, statistics.VacationHours.Length);
+        Assert.Equal(2, statistics.InvoiceRate.Length);
+        Assert.Equal(2, statistics.NonBillableInvoiceRate.Length);
+        Assert.Equal(2, statistics.Start.Length);
+        Assert.Equal(2, statistics.End.Length);
 
         Assert.Equal(1m, statistics.InvoiceRate[0]);
         Assert.Equal(new DateTime(2022, 1, 1), statistics.Start[0]);
@@ -516,10 +517,10 @@ public class InvoiceRateServiceTest
         var endDate = new DateTime(2023, 12, 31);
 
         var hours = new List<Hours>();
-        for (int i = 0; i < 730; i++)
+        for (var i = 0; i < 730; i++)
         {
             var newDate = startDate.AddDays(i);
-            if (newDate.DayOfWeek == DayOfWeek.Sunday || newDate.DayOfWeek == DayOfWeek.Saturday)
+            if (newDate.DayOfWeek is DayOfWeek.Sunday or DayOfWeek.Saturday)
             {
                 continue;
             }
@@ -540,13 +541,13 @@ public class InvoiceRateServiceTest
 
         var statistics = await service.GetEmployeeInvoiceStatisticsByPeriod(startDate, endDate, InvoicePeriods.Annualy, ExtendPeriod.None, false);
 
-        Assert.Equal(2, statistics.BillableHours.Count());
-        Assert.Equal(2, statistics.NonBillableHours.Count());
-        Assert.Equal(2, statistics.VacationHours.Count());
-        Assert.Equal(2, statistics.InvoiceRate.Count());
-        Assert.Equal(2, statistics.NonBillableInvoiceRate.Count());
-        Assert.Equal(2, statistics.Start.Count());
-        Assert.Equal(2, statistics.End.Count());
+        Assert.Equal(2, statistics.BillableHours.Length);
+        Assert.Equal(2, statistics.NonBillableHours.Length);
+        Assert.Equal(2, statistics.VacationHours.Length);
+        Assert.Equal(2, statistics.InvoiceRate.Length);
+        Assert.Equal(2, statistics.NonBillableInvoiceRate.Length);
+        Assert.Equal(2, statistics.Start.Length);
+        Assert.Equal(2, statistics.End.Length);
 
         Assert.True(1m < statistics.InvoiceRate[0]);
         Assert.Equal(startDate, statistics.Start[0]);
